making prgress
