+++ Snake-AI/Globals.hpp
@@ -1,7 +1,12 @@
 #pragma once
 
+#include <vector>
+
 constexpr int WINDOW_SIZE = 800;
 constexpr int PIXEL_SIZE = 20;
 constexpr int BLOCK_SIZE = WINDOW_SIZE / PIXEL_SIZE;
 constexpr int FPS_LIMIT = 60;
 constexpr float SNAKE_SPEED = 7.f;
+
+bool isError = false;
+std::vector<std::string> Errors;
