+++ Snake-AI/Snake-AI.vcxproj
@@ -146,15 +146,18 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
-    <ClCompile Include="main.cpp" />
+    <ClCompile Include="Food.cpp" />
+    <ClCompile Include="Main.cpp" />
+    <ClCompile Include="Snake.cpp" />
     <ClCompile Include="SnakeGame.cpp" />
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="DrawObjects.h" />
-    <ClInclude Include="Globals.h" />
-    <ClInclude Include="resource.h" />
-    <ClInclude Include="Snake.h" />
-    <ClInclude Include="SnakeGame.h" />
+    <ClInclude Include="DrawObjects.hpp" />
+    <ClInclude Include="Food.hpp" />
+    <ClInclude Include="Globals.hpp" />
+    <ClInclude Include="resource.hpp" />
+    <ClInclude Include="Snake.hpp" />
+    <ClInclude Include="SnakeGame.hpp" />
   </ItemGroup>
   <ItemGroup>
     <Font Include="font.ttf" />
