+++ Snake-AI/Snake.cpp
@@ -1,0 +1,191 @@
+#include "Snake.hpp"
+#include "Globals.hpp"
+
+Snake::Snake(const sf::Vector2f& pos) {
+    positions.push_back(pos);
+}
+
+void Snake::move() {
+    //inserts new pos to positions depending on the direction it is moving
+    positions.insert(positions.begin(), sf::Vector2f(positions[0].x + changeX, positions[0].y + changeY));
+}
+
+void Snake::setTextures() {
+    if (!tail_left.loadFromFile("positions_texture/tail_left.png")) {
+        isError = true;
+    }
+    else {
+        tailLeft.setTexture(tail_left);
+    }
+
+    sf::Texture tail_right; // ->
+    if (!tail_right.loadFromFile("positions_texture/tail_right.png")) {
+        isError = true;
+    }
+    else {
+        tailRight.setTexture(tail_right);
+    }
+
+    sf::Texture tail_down; // V
+    if (!tail_down.loadFromFile("positions_texture/tail_down.png")) {
+        isError = true;
+    }
+    else {
+        tailDown.setTexture(tail_down);
+    }
+
+    sf::Texture tail_up; // A
+    if (!tail_up.loadFromFile("positions_texture/tail_up.png")) {
+        isError = true;
+    }
+    else {
+        tailUp.setTexture(tail_up);
+    }
+
+    sf::Texture horizontal; // --
+    if (!horizontal.loadFromFile("positions_texture/body_horizontal.png")) {
+        isError = true;
+    }
+    else {
+        horiz.setTexture(horizontal);
+    }
+
+    sf::Texture vertical; // |
+    if (!vertical.loadFromFile("positions_texture/body_vertical.png")) {
+        isError = true;
+    }
+    else {
+        vertic.setTexture(vertical);
+    }
+
+    sf::Texture topright; // L
+    if (!topright.loadFromFile("positions_texture/body_topright.png")) {
+        isError = true;
+    }
+    else {
+        topRight.setTexture(topright);
+    }
+
+    sf::Texture topleft; // _|
+    if (!topleft.loadFromFile("positions_texture/body_topleft.png")) {
+        isError = true;
+    }
+    else {
+        topLeft.setTexture(topleft);
+    }
+
+    sf::Texture bottomleft; // -,
+    if (!bottomleft.loadFromFile("positions_texture/body_bottomleft.png")) {
+        isError = true;
+    }
+    else {
+        bottomLeft.setTexture(bottomleft);
+    }
+
+    sf::Texture bottomright; // F
+    if (!bottomright.loadFromFile("positions_texture/body_bottomright.png")) {
+        isError = true;
+    }
+    else {
+        bottomRight.setTexture(bottomright);
+    }
+
+    sf::Texture head_left; // <- hlava
+    if (!head_left.loadFromFile("positions_texture/head_left.png")) {
+        isError = true;
+    }
+    else {
+        headLeft.setTexture(head_left);
+    }
+
+    sf::Texture head_right; // -> hlava
+    if (!head_right.loadFromFile("positions_texture/head_right.png")) {
+        isError = true;
+    }
+    else {
+        headRight.setTexture(head_right);
+    }
+
+    sf::Texture head_down; // V hlava
+    if (!head_down.loadFromFile("positions_texture/head_down.png")) {
+        isError = true;
+    }
+    else {
+        headDown.setTexture(head_down);
+    }
+
+    sf::Texture head_up; // A hlava
+    if (!head_up.loadFromFile("positions_texture/head_up.png")) {
+        isError = true;
+    }
+    else {
+        headUp.setTexture(head_up);
+    }
+}
+
+void Snake::draw(sf::RenderWindow& window) {
+    if (changeX == -BLOCK_SIZE) { //head
+        headLeft.setPosition(positions[0]);
+        window.draw(headLeft);
+    }
+    else if (changeX == BLOCK_SIZE) {
+        headRight.setPosition(positions[0]);
+        window.draw(headRight);
+    }
+    else if (changeY == -BLOCK_SIZE) {
+        headUp.setPosition(positions[0]);
+        window.draw(headUp);
+    }
+    else if (changeY == BLOCK_SIZE) {
+        headDown.setPosition(positions[0]);
+        window.draw(headDown);
+    }
+
+    if (positions.size() > 1) { //tail
+        if (positions[positions.size() - 2].x == positions[positions.size() - 1].x && positions[positions.size() - 2].y > positions[positions.size() - 1].y) {
+            tailUp.setPosition(positions[positions.size() - 1]);
+            window.draw(tailUp);
+        }
+        else if (positions[positions.size() - 2].x == positions[positions.size() - 1].x && positions[positions.size() - 2].y < positions[positions.size() - 1].y) {
+            tailDown.setPosition(positions[positions.size() - 1]);
+            window.draw(tailDown);
+        }
+        else if (positions[positions.size() - 2].x < positions[positions.size() - 1].x && positions[positions.size() - 2].y == positions[positions.size() - 1].y) {
+            tailRight.setPosition(positions[positions.size() - 1]);
+            window.draw(tailRight);
+        }
+        else if (positions[positions.size() - 2].x > positions[positions.size() - 1].x && positions[positions.size() - 2].y == positions[positions.size() - 1].y) {
+            tailLeft.setPosition(positions[positions.size() - 1]);
+            window.draw(tailLeft);
+        }
+    }
+
+    if (positions.size() > 2) { // if positionss length is bigger then 2 it has middle positions part (chest or smth like that) (everything except tail and head)
+        for (int i = 1; i < positions.size() - 1; i++) {
+            if (positions[i + 1].x == positions[i - 1].x && positions[i + 1].y != positions[i - 1].y) {
+                vertic.setPosition(positions[i]);
+                window.draw(vertic);
+            }
+            else if (positions[i + 1].x != positions[i - 1].x && positions[i + 1].y == positions[i - 1].y) {
+                horiz.setPosition(positions[i]);
+                window.draw(horiz);
+            }
+            else if ((positions[i - 1].x > positions[i].x && positions[i + 1].y > positions[i].y) || (positions[i + 1].x > positions[i].x && positions[i - 1].y > positions[i].y)) {
+                bottomRight.setPosition(positions[i]);
+                window.draw(bottomRight);
+            }
+            else if ((positions[i - 1].x < positions[i].x && positions[i + 1].y < positions[i].y) || (positions[i + 1].x < positions[i].x && positions[i - 1].y < positions[i].y)) {
+                topLeft.setPosition(positions[i]);
+                window.draw(topLeft);
+            }
+            else if ((positions[i - 1].x < positions[i].x && positions[i + 1].y > positions[i].y) || (positions[i + 1].x < positions[i].x && positions[i - 1].y > positions[i].y)) {
+                bottomLeft.setPosition(positions[i]);
+                window.draw(bottomLeft);
+            }
+            else if ((positions[i - 1].x > positions[i].x && positions[i + 1].y < positions[i].y) || (positions[i + 1].x > positions[i].x && positions[i - 1].y < positions[i].y)) {
+                topRight.setPosition(positions[i]);
+                window.draw(topRight);
+            }
+        }
+    }
+}
